scripts/requirements.txt
beautifulsoup4>=4.12
requests>=2.31
pyyaml>=6.0
duckdb>=1.0
